\name{fpgenomat}
\alias{fpgenomat}
\docType{data}
\title{genotype matrix for the two parents
}
\description{
genotype matrix for the two parents. 
}
\usage{data(fpgenomat)}
\format{
  A data frame or matrix with 2*nmar observations.column names should be marker names.
}

\source{
Monsanto breeding data
}
\examples{
data(fpgenomat);
str(fpgenomat) ; 
}
