\name{parentsGeno}
\alias{parentsGeno}
\docType{data}
\title{genotype matrix for the two parents
}
\description{
genotype matrix for the two parents. 
}
\usage{
data(parentsGeno)
}
\format{
  A data frame or matrix with 2 rows: each row represent the marker genotypes of one parent.
}


\examples{
data(parentsGeno);
str(parentsGeno) ; 
}
